requirements.txt
pywebview
spotipy
google-generativeai
python-dotenv
